requirements.txt
python-dotenv
azure-core
azure-identity
azure-ai-projects
openai
azure-ai-textanalytics==5.3.0